% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{allObs.Rendata}
\alias{allObs.Rendata}
\title{Retrieve All Observations of a \code{Rendata} object as Rows of a single
Data Frame}
\usage{
allObs.Rendata(object, ...)
}
\arguments{
\item{object}{An object with (S3) class \code{"Rendata"} from the
\pkg{Renext} package.}

\item{...}{Not used yet.}
}
\value{
A data frame that can be used with the \code{ggplot}
    function. This data frame has the four columns \code{x},
    \code{xend}, \code{y} and \code{yend} matching in the
    aesthetic of a \code{\link{geom_segment}} layer. It also has
    the columns \code{type} (with values \code{"OT"}, \code{"MAX"}
    and \code{"OTS"}) and \code{block} which identifies the
    block. Mind that \code{x} and \code{xend} are objects with
    class \code{"Date"} or \code{"POSIXct"}.
}
\description{
Gather the different types of observations of \code{object} into a
single data frame that can be used to build a \code{ggplot}.
}
\details{
The \code{"OT"} observations will be displayed as vertical
segments, while the \code{"MAX"} or \code{"OTS"} may be displayed
as horizontal segments when their date is not known. The columns
\code{x}, \code{xend}, \code{y} and \code{yend} are defined on
purpose.
}
\examples{
rd3 <- rRendata(effDuration = 10,
                OTS.effDuration = rep(60, 2),
                OTS.threshold = rep(80, 2),
                simDate = FALSE,
                distname.y = "gpd", par.y = c(scale = 20, shape = 0.16))
df <- allObs.Rendata(rd3)
g <- ggplot(data = df)
g <- g + geom_segment(mapping = aes(x = x, xend = xend,
                                    y = y, yend = yend, colour = type))
g

}
\keyword{internal}
