% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{allObs.Renouv}
\alias{allObs.Renouv}
\title{Retrieve All Observations of a \code{Rendata} object as
    Rows of a single Data Frame}
\usage{
allObs.Renouv(object, byBlockStyle = NULL, posOptions = NULL, ...)
}
\arguments{
\item{object}{An object with class \code{"Rendata"}.}

\item{byBlockStyle}{A named list with logical elements.}

\item{posOptions}{A list of options for the computations of the
plotting positions i.e. of optional arguments to be passed to
\code{\link[Renext]{SandT}}.}

\item{...}{Not used yet.}
}
\value{
A list of two data frames that can be used with the
    \code{ggplot} function.

\describe{
   \item{\code{dfST} }{

       This data frame is to be used in a \code{geom_point} layer.
       It has variables \code{Period}, \code{Quantile} and
       \code{Group} providing \code{x}, \code{x} and \code{group} or
       \code{colour} in the aesthetic.

   }
   \item{\code{dfSeg} }{ 

       Has variables \code{x}, \code{xend}, \code{y} and
       \code{yend} related to the aesthetic of a
       \code{\link{geom_segment}} layer. The column \code{Group}
       identifies the group. \bold{Caution}: can have zero rows.

   }
}
}
\description{
Gather the different types of observations described in the
\code{Renouv} object \code{object} and related \emph{plotting
positions} into a single data frame that can be used to build a
\code{ggplot}.
}
\details{
For each observation, a plotting position is provided to show the
observation as a point on the \emph{return level plot}. A special
case concerns \code{"OTS"} blocks with no exceedance over the
threshold hence with no observation. Such a block will be shown by
using a horizontal segment ending at a suitable return-period.
}
\examples{
example(Renouv, ask = FALSE, echo = FALSE)
res <- allObs.Renouv(fit3)

}
\keyword{internal}
