% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{allPeriods.Rendata}
\alias{allPeriods.Rendata}
\title{Periods in a \code{Rendata} Object}
\usage{
allPeriods.Rendata(object, ...)
}
\arguments{
\item{object}{An object with class \code{"Rendata"}.}

\item{...}{Not used yet.}
}
\value{
A data frame that can be used with the \code{ggplot} function.
}
\description{
Gather all the periods described in \code{object} into a single
data frame that can be used to build a \code{ggplot}.
}
\examples{
rd3 <- rRendata(effDuration = 10,
                OTS.effDuration = rep(60, 2),
                OTS.threshold = rep(80, 2),
                simDate = FALSE,
                distname.y = "gpd", par.y = c(scale = 20, shape = 0.16))
df <- allPeriods.Rendata(Brest)
g <- ggplot(data = df)
g <- g + geom_rect(mapping = aes(xmin = start, xmax = end, ymin = 0, ymax = 1, colour = period,
                                 fill = period))
g
}
\keyword{internal}
