% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotRenouvTList.R
\name{autoplot.coef.RenouvTList}
\alias{autoplot.coef.RenouvTList}
\title{Build a \code{ggplot} for the Coeficients of a
    \code{RenouvTList} Object}
\usage{
\method{autoplot}{coef.RenouvTList}(object, facets, ...)
}
\arguments{
\item{object}{A \code{RenouvTList} object}

\item{facets}{Logical. It \code{TRUE} each parameter is shown in
one facet oa trellis graphics.}

\item{...}{Not used yet.}
}
\value{
A \code{ggplot} object.
}
\description{
Build a \code{ggplot} for the Coeficients of a
    \code{RenouvTList} Object
}
\examples{
fit <- RenouvTList(Garonne,
                   threshold = seq(from = 2401, to = 3001, by = 10),
                   distname.y = "GPD")
fitJit <- RenouvTList(GaronneJit,
                      threshold = seq(from = 2401, to = 3001, by = 10),
                      distname.y = "GPD")
autoplot(coef(fit))
autoplot(coef(fitJit))

}
