% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotRenouvTList.R
\name{autoplot.coSd.RenouvTList}
\alias{autoplot.coSd.RenouvTList}
\title{Generate a \code{ggplot} from a \code{coSd.RenouvTList} Object}
\usage{
\method{autoplot}{coSd.RenouvTList}(object, facets, ...)
}
\arguments{
\item{object}{An object with class \code{"coSd.RenouvTlist"} as
created by applying the \code{coSd} method to a
\code{RenouvTlist} object.}

\item{facets}{Logical. If \code{TRUE} the created \code{ggplot}
object is a trellis graphics with one facet for each threshold
value.}

\item{...}{Not used yet.}
}
\description{
The estimates of the POT parameters are plotted
    against the threshold. These plots are sometimes called
    "threshold stability plots".
}
\examples{
fitJit <- RenouvTList(GaronneJit,
                      threshold = seq(from = 2401, to = 3001, by = 100),
                      distname.y = "GPD")
cs <- coef(fitJit, sd = TRUE)
autoplot(cs)


}
