% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotRendata.R
\name{autoplot.Rendata}
\alias{autoplot.Rendata}
\title{Generate a \code{ggplot} from a \code{Rendata} Object.}
\usage{
\method{autoplot}{Rendata}(
  object,
  textover = quantile(object$OTdata[, object$info$varName], probs = 0.99),
  fillPeriods = TRUE,
  showHist = TRUE,
  ...
)
}
\arguments{
\item{object}{An object with class \code{"Rendata"} as defined by.
\pkg{Renext} package.}

\item{textover}{The \code{"OT"} observations exceeding this value
will be marked with their date.}

\item{fillPeriods}{Logical. If \code{TRUE} the periods are shown
using semi-transparent filled rectangles.}

\item{showHist}{Logical. If \code{TRUE} the "historial" periods ("OTS" and
"MAX") and their observations will be shown.}

\item{...}{Not used yet.}
}
\value{
An object inheriting from \code{"ggplot"}.
}
\description{
Generate a \code{ggplot} from a \code{Rendata} Object.
}
\details{
The "historial" periods ("OTS" and "MAX") and their observations
may be shown or not. Mind that an \code{"OTS"} period may have no
observations so it is a good idea to used \code{fillPeriods =
TRUE} when \code{showHist} is \code{TRUE}.
}
\examples{
autoplot(Garonne)
set.seed(1234)
rd <- rRendata(effDuration = 60,
               MAX.effDuration = rep(3, 6), MAX.r = rep(4, 6),
               distname.y = "exp", par.y = c(rate = 1/100))
autoplot(rd)
rd3 <- rRendata(effDuration = 10,
                OTS.effDuration = rep(60, 2),
                OTS.threshold = rep(80, 2),
                simDate = FALSE,
                distname.y = "gpd", par.y = c(scale = 20, shape = 0.16))
autoplot(rd3)

}
