% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotRenouv.R
\name{autoplot.Renouv}
\alias{autoplot.Renouv}
\title{Generate a \code{ggplot} from a \code{Renouv} Object}
\usage{
\method{autoplot}{Renouv}(
  object,
  level = 0.95,
  show = list(quant = TRUE, conf = TRUE, allObs = TRUE),
  posOptions = NULL,
  byBlockStyle = NULL,
  ...
)
}
\arguments{
\item{object}{An object with class \code{"Renouv"} representing a
fitted Marked Poisson process model.}

\item{level}{The confidence level.}

\item{show}{A named list describing the elements to be shown. A
missing information is equivalent to \code{FALSE}. The value
of the element \code{"quant"} tells if the quantile (or return
level) line is to be shown or not. The element \code{"conf"}
is for the confidence interval(s) and the element
\code{"allObs"} is for the observations.}

\item{posOptions}{A named list of arguments to be passed to the
\code{\link{allObs.Renouv}} and then to the
\code{\link[Renext]{SandT}} function. Mind that in order to be
consistent with the \code{\link[Renext]{plot.Renouv}} method
of \pkg{Renext}, the "p-points" are used by default rather the
"H-points" (a.k.a. Nelson's points). Also the default choice
is corresponds to the so-called \emph{Weibull positions}
corresponding to \code{a = 0} in \code{\link[stats]{ppoints}}
and these positions are biassed.}

\item{byBlockStyle}{As in \code{\link{plot.Renouv}}.}

\item{...}{Not used yet.}
}
\description{
Generate a \code{ggplot} from a \code{Renouv} object. The plot
shows the return-level curve using a log-scale for the return
period.  It also displays the (possibly censored) observations
with suitable plotting positions, and confidence intervals on the
return levels.
}
\details{
Although this method can be viewed as as substitute of
    \code{\link[Renext]{plot.Renouv}} method of \pkg{Renext}, some
    differences in the appearence result from the use of
    \pkg{ggplot2} instead of \pkg{graphics}. The content of the
    legend(s) can be quite different from that of the single
    legend obtained with \code{plot.Renouv}. Note that one can
    choose to show or not the observations, but the choice is for
    \emph{all the observations}: \code{OT} and the historical
    observations \code{OTS} and \code{MAX} if they exist.
}
\examples{
example(Renouv, ask = FALSE, echo = FALSE)
autoplot(fit3)
autoplot(fit3, byBlockStyle = list(OTS = FALSE))
autoplot(fit3, level = c(0.7, 0.90))
}
