% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenouvTList.R
\name{coef.RenouvTList}
\alias{coef.RenouvTList}
\title{Coefficients of a \code{RenouvTList} Object}
\usage{
\method{coef}{RenouvTList}(object, reParam = TRUE, ...)
}
\arguments{
\item{object}{An object with class \code{"RenouvTList"} as created
by using \code{\link{RenouvTList}}.}

\item{...}{Not used yet.}
}
\value{
A numeric matrix with its rows corresponding to the
    thresholds.
}
\description{
Coefficients of a \code{RenouvTList} Object
}
\examples{
fit <- RenouvTList(Garonne,
                   threshold = seq(from = 2401, to = 3001, by = 10),
                   distname.y = "GPD")
autoplot(coef(fit))

}
