% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenouvTList.R
\name{coSd.RenouvTList}
\alias{coSd.RenouvTList}
\title{Estimated Coefficients and their Standard Deviation}
\usage{
\method{coSd}{RenouvTList}(object, reParam = TRUE, ...)
}
\arguments{
\item{object}{A \code{RenouvTList} object.}

\item{reParam}{Logical. Can only be used when the distribution is
\code{"GPD"}. If \code{TRUE} the GPD scale parameter
\eqn{\sigma} is replaced by the modified scale
\eqn{\sigma^\star := \sigma - u \xi}, where \eqn{u} is the
threshold and \eqn{\xi} is the GPD shape. The parameter
\eqn{\sigma^\star} does not depend on the threshold. Remind
that the distribution is specified by using the
\code{distname.y} argument of the creator \code{RenouvTList}.}

\item{...}{Not used yet.}
}
\description{
Extract the estimated coefficients and their standard
    deviation as an object with class \code{"coSd.RenouvTList"} for
    which some methods are available such as \code{autoplot} and
    \code{print}.
}
\details{
The POT parameters of a \code{Renouv} object with a GP
    distribution of the eceedances are named \code{"lambda"},
    \code{"scale"} and \code{"shape"}. If a reparameterization is
    used \code{reParam = TRUE}, the modified GP scale
    \eqn{\sigma^\star} is named \code{scale.ind} to recall the
    independence from the threshold. Note that although the
    \code{print} method displays the object as a matrix, it
    actually consists in a list of two matrices.
}
\references{
Coles S. (2001). \emph{An Introduction to Statistical
    Modeling of Extreme Values} Springer-Verlag.
}
