% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGPD3.R
\name{dEGPD3}
\alias{dEGPD3}
\alias{pEGPD3}
\alias{qEGPD3}
\alias{rEGPD3}
\title{Extended Generalised Pareto Distribution, Type 3}
\usage{
dEGPD3(x, scale = 1, shape = 0, kappa = 1, log = FALSE)

pEGPD3(q, scale = 1, shape = 0, kappa = 1, lower.tail = TRUE)

qEGPD3(p, scale = 1, shape = 0, kappa = 1)

rEGPD3(n = 1L, scale = 1, shape = 0, kappa = 1)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{scale, shape}{GPD-like scale and shape parameters.}

\item{kappa}{Shape coefficient \eqn{\kappa}.}

\item{log}{Logical. If \code{TRUE} the log-density is returned.}

\item{q}{Vector of quantiles.}

\item{lower.tail}{Logical. If \code{TRUE} (default) the
probabilities are \eqn{\textrm{Pr}[X \leq x]}{Pr[X <= x]} and
\eqn{\textrm{Pr}[X > x]}{Pr[X > x]} otherwise.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
\code{dEGPD3} is the density function, \code{pEGPD3} is
the distribution function, \code{qEGPD3} is the quantile function
and \code{rEGPD3} generates random variables.
}
\description{
The Extended Generalised Pareto Distribution type 3 was introduced
by Papastathopoulos and Tawn in 2013, along with two other
distributions built in a similar way. These distributions are
intended to be used in the Peak Over Threshold framework with a
reduced impact of the threshold choice.
}
\details{
The distribution depends on the three parameters: \emph{scale}
\eqn{\sigma >0}, \emph{shape} \eqn{\xi} and a "non-tail" parameter
\eqn{\kappa >0}. Its distribution function
\eqn{G_{\text{EGPD3}}(x;\,\sigma,\,\xi,\,\kappa)}{G(x; \sigma, \xi, \kappa)}
relates to the GPD
\eqn{F_{\text{GPD}}(x;\,\sigma,\,\xi)}{F(x; \sigma, \xi, \kappa)}
according to
\deqn{F_{\text{EGPD3}}(x; \sigma,\,\xi,\,\kappa) =
F_{\text{GPD}}(x; \,\sigma,\,\xi)^\kappa}{G(x, \sigma, \xi,
\kappa) = F(x, \sigma, \xi)^\kappa} The distribution is
tail-equivalent to the two-parameter GPD with the same scale
\eqn{\sigma} and shape \eqn{\xi}.
}
\examples{
xi <- 0.2
kappa <- 1.1
x <- qEGPD3(c(0.001, 0.999), shape = xi, kappa = kappa)
x <- seq(from = x[1], to = x[2], length.out = 300)
plot(x, dEGPD3(x, shape = xi, kappa = kappa), type = "l",
     xlab = "", ylab = "dens", main = "EGPD3 density")

fit <- RenouvTList(x = GaronneJit, distname.y = "GPD",
                   threshold = seq(from = 2401, to = 3001, by = 100),
                   start.par.y = c(scale = 2000, shape = -0.2, kappa = 1.0))
autoplot(predict(fit))
## Now use the 'EGPD3' distribution for the exceedances.
\dontrun{
fitE <- RenouvTList(x = GaronneJit, distname.y = "EGPD3",
                   threshold = seq(from = 2401, to = 3001, by = 100),
                   start.par.y = c(scale = 2000, shape = -0.2, kappa = 1.0))
coef(fitE, reParam = FALSE)
autoplot(predict(fitE))
}
}
\references{
Iaonnis Papastathopoulos and Jonathan A. Tawn (2013).
"Extended generalised Pareto models for tail estimation".
\emph{J. Stat. Plan. Inference}, \bold{143}(3): 131-143,
\doi{https://doi.org/10.1016/j.jspi.2012.07.001}.
}
