% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenouvTList.R
\name{predict.RenouvTList}
\alias{predict.RenouvTList}
\title{Predict Method for \code{RenouvTList} Objects}
\usage{
\method{predict}{RenouvTList}(object, newdata, level = 0.95, ...)
}
\arguments{
\item{object}{An object with class \code{"RenouvTList"}.}

\item{...}{Further arguments to be passed to \code{predict}.}
}
\value{
An object with class \code{"predict.RenouvTList"}
    inheriting from \code{"data.frame"}
}
\description{
Compute a data frame of return levels for the all the
    \code{Renouv} objects stored in \code{object}.
}
\details{
The returned object is given a special class in order to
    allow the use of some S3 methods such as \code{autoplot}.
}
