% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenouvTList.R
\name{RenouvTList}
\alias{RenouvTList}
\title{\code{Renouv} or POT Models with several Thresholds}
\usage{
RenouvTList(x, threshold = NULL, effDuration = NULL, distname.y = "GPD", ...)
}
\arguments{
\item{x}{A numeric vector or a suitable \code{Rendata} object, see
\code{\link{Renouv}}.}

\item{threshold}{A numeric vector of thresholds.}

\item{effDuration, distname.y}{Effective duration and distribution
See \code{\link{Renouv}}.}

\item{...}{Further arguments to be passed to
\code{\link{Renouv}}. Graphical arguments should not be used.}

\item{loopIni}{Logical. If \code{TRUE} the initial values are This
is because no convergence problems should arise as long as no
historical/censored observations are used.}
}
\value{
An object with (S3) class \code{"RenouvTList"}. This S3
    class has a number of methods.
}
\description{
Fit a \code{Renouv} or POT model for several
    candidate thresholds. The result is an object having the S3
    class \code{"RenouvTList"} for which several methods have been
    implemented in \pkg{Renextra}.
}
\details{
Although the classical methods like \code{coef} are
    mainly obtained with \code{lapply} or \code{sapply}, the
    result is in some cases given a special class that allows
    further invetigation. See the \bold{Example} section.
}
\section{Caution}{
 the default distribution for the exceedances is
    taken to be the usual two-parameter Generalised Pareto
    \code{"GPD"} rather than exponential.
}

\examples{
fit <- RenouvTList(Garonne,
                   threshold = seq(from = 2401, to = 3001, by = 100),
                   distname.y = "GPD")

methods(class = "RenouvTList")
autoplot(predict(fit))
fitJit <- RenouvTList(GaronneJit,
                      threshold = seq(from = 2401, to = 3001, by = 100),
                      distname.y = "GPD")
autoplot(predict(fitJit))

}
