% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplotRenouv.R
\name{autoplot.Renouv}
\alias{autoplot.Renouv}
\title{Generate a \code{ggplot} from a \code{Renouv} Object}
\usage{
\method{autoplot}{Renouv}(
  object,
  level = 0.95,
  show = list(OT = TRUE, quant = TRUE, conf = TRUE, MAX = TRUE, OTS = TRUE),
  posOptions = NULL,
  byBlockStyle = NULL,
  ...
)
}
\arguments{
\item{object}{An object with class \code{"Renouv"} representing a
fitted Marked Poisson process model.}

\item{level}{The confidence level.}

\item{show}{A named list describing the elements to be shown. A
missing information is equivalent to FALSE.}

\item{posOptions}{A named list of arguments to be passed to the
\code{\link{SandT}} function. Mind that in order to be
consistent with the \code{plot.Renouv} method, the "p-points"
are used rather the "H-points" (or Nelson's points).}

\item{byBlockStyle}{As in \code{\link{plot.Renouv}}.}

\item{...}{Not used yet.}
}
\description{
Generate a \code{ggplot} from a \code{Renouv} object. The plot
shows the return-level curve using a log-scale for the return
period.  It also displays the (possibly censored) observations
with suitable plotting positions, and confidence intervals on the
return levels.
}
\details{
Although this can be viewed as as substitute of
    \code{plot.Renouv} method of \pkg{Renext}, some differences in
    the appearence result from the use of \pkg{ggplot2} instead of
    \pkg{graphics}. The content of the legend(s) can be quite
    different from that of the single legend obtained with
    \code{plot.Renouv}.
}
\examples{
example(Renouv, ask = FALSE, echo = FALSE)
autoplot(fit3) 
}
