% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenouvTList.R
\name{coef.RenouvTList}
\alias{coef.RenouvTList}
\title{Coefficients of a \code{RenouvTList} Object}
\usage{
\method{coef}{RenouvTList}(object, reParam, lambda = TRUE, sd = TRUE, ...)
}
\arguments{
\item{object}{An object with class \code{"RenouvTList"} as created
by using \code{\link{RenouvTList}}.}

\item{reParam}{Logical. Can only be used when the distribution is
\code{"GPD"}. If \code{TRUE} the GPD scale parameter
\eqn{\sigma} is replaced by the modified scale
\eqn{\sigma^\star := \sigma - u \xi}, where \eqn{u} is the
threshold and \eqn{\xi} is the GPD shape. The parameter
\eqn{\sigma^\star} does not depend on the threshold. Remind
that the distribution is specified by using the
\code{distname.y} argument of the creator \code{RenouvTList}.}

\item{lambda}{Logical. If \code{TRUE} the Poisson rate is included
in the results. The choice \code{FALSE} is relevant for
threshold stability analyzes.}

\item{sd}{Logical. If \code{TRUE} the result is a list of
\emph{two} matrices: one for the estimates and one for the
standard deviations of these, a.k.a. the \emph{standard
errors}.}

\item{...}{Not used yet.}
}
\value{
A numeric matrix with its rows corresponding to the
    thresholds.
}
\description{
Coefficients of a \code{RenouvTList} Object
}
\section{Caution}{
 the confidence intervals are obtained by using
    the "delta" method. The intervals are known to have a coverage
    rate which is smaller than the expected rate when the
    Generalised Pareto Distribution is used.
}

\examples{
fit <- RenouvTList(Garonne,
                   threshold = seq(from = 2401, to = 3001, by = 10),
                   distname.y = "GPD")
autoplot(coef(fit))

}
